clang/mac: Fix most easy problems in chrome.xcodeproj (all targets)
BUG=None
TEST=None

Review URL: http://codereview.chromium.org/3550006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@61280 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e3a8a926da1f7e3365a6f0dbb9061af975a36361

+++ chrome/browser/cocoa/accelerators_cocoa_unittest.mm
@@ -17,7 +17,7 @@ TEST(AcceleratorsCocoaTest, GetAccelerator) {
       keymap->GetAcceleratorForCommand(IDC_COPY);
   ASSERT_TRUE(accelerator);
   EXPECT_NSEQ(@"c", accelerator->characters());
-  EXPECT_EQ(NSCommandKeyMask, accelerator->modifiers());
+  EXPECT_EQ(static_cast<int>(NSCommandKeyMask), accelerator->modifiers());
 }
 
 TEST(AcceleratorsCocoaTest, GetNullAccelerator) {
